Moved 1715 to the ApiGuidelinesAnalyzer. Fixed up tests to test the message with the name of the identifier as well.
