+++ src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/Microsoft.ApiDesignGuidelines.CSharp.Analyzers.csproj
@@ -124,7 +124,6 @@
     <Compile Include="CSharpStaticHolderTypesShouldNotHaveConstructors.cs" />
     <Compile Include="CSharpStaticHolderTypesShouldNotHaveConstructors.Fixer.cs" />
     <Compile Include="CSharpMovePInvokesToNativeMethodsClass.Fixer.cs" />
-    <Compile Include="CSharpIdentifiersShouldHaveCorrectPrefix.cs" />
     <Compile Include="CSharpIdentifiersShouldHaveCorrectPrefix.Fixer.cs" />
     <Compile Include="CSharpNonConstantFieldsShouldNotBeVisible.cs" />
     <Compile Include="CSharpNonConstantFieldsShouldNotBeVisible.Fixer.cs" />
